Function to create a template for the multistage design used in tmt

This function creates a template for all multistage design related inputs for the tmt package (tmt_mstdesign). Essentially, this are the modules, rules and path sections. In the formula-based notation, it is also possible to state additional conditions (constraints) that can be found in the data and are reflected in the multistage design.

create_mstdesign(formula, full = TRUE, eval = TRUE)

Arguments

formula
formula for the desired template of a multistage design

full
logical if the modules and rules sections should also be created

eval
logical should the text input be evaluated (e.g. 3:6 = c(3, 4, 5, 6))

Author

Jan Steinfeld

Examples

#############################################################################
# create simple template
#############################################################################
formula = "start(start) += S1(B1,B2,B3) += S2(B4,B5,B6,B7)"
create_mstdesign(formula, full = TRUE, eval = TRUE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> start( ) += B1( ) += B4( ) 
#> start( ) += B2( ) += B4( ) 
#> start( ) += B3( ) += B4( ) 
#> start( ) += B1( ) += B5( ) 
#> start( ) += B2( ) += B5( ) 
#> start( ) += B3( ) += B5( ) 
#> start( ) += B1( ) += B6( ) 
#> start( ) += B2( ) += B6( ) 
#> start( ) += B3( ) += B6( ) 
#> start( ) += B1( ) += B7( ) 
#> start( ) += B2( ) += B7( ) 
#> start( ) += B3( ) += B7( ) 
create_mstdesign(formula, full = TRUE, eval = FALSE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> start( ) += B1( ) += B4( ) 
#> start( ) += B2( ) += B4( ) 
#> start( ) += B3( ) += B4( ) 
#> start( ) += B1( ) += B5( ) 
#> start( ) += B2( ) += B5( ) 
#> start( ) += B3( ) += B5( ) 
#> start( ) += B1( ) += B6( ) 
#> start( ) += B2( ) += B6( ) 
#> start( ) += B3( ) += B6( ) 
#> start( ) += B1( ) += B7( ) 
#> start( ) += B2( ) += B7( ) 
#> start( ) += B3( ) += B7( ) 

#############################################################################
# create complex template
#############################################################################

formula = "nativ(no,yes) ~ education(low,medium,heigh) ~ 
            CBM(3:6) += S1(B1,B2,B3) += S2(B4,B5,B6,B7)"
create_mstdesign(formula, full = TRUE, eval = TRUE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B7( ) 
create_mstdesign(formula, full = TRUE, eval = FALSE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B7( )