Computation of Andersen's Likelihood-Ratio Test

This function applies the Likelihood Ratio Test of Andersen. Note that all persons with raw score equal to "median" are assigned to lower group. Is is also allowed to split after "mean" or submit any dichotomous vector.

tmt_lrtest(object, split = "median", cores = NULL, se = TRUE, ...)

Arguments

object
it is necessary to submit an object of the function mst or nmst

split
default is the split criteria "median" of the raw score, optional are "mean" or any dichotomouse vector

cores
submit integer of cores you would like to apply

se
logical: if true, the standard error is estimated

…
further arguments for the tmt_rm function

Value

List with following entries

data_orig

Submitted data frame with item responses

betapars_subgroup

List of item parameters (difficulty) for each subgroup

se.beta_subgroup

List of standard errors of the estimated item parameters

model

Used model ((mst) for Rasch model with multistage design)

LRvalue

LR-value

df

Degrees of freedoms for the test statistic

pvalue

P-value of the likelihood ratio test

loglik_subgroup

Log-likelihoods for the subgroups

split_subgroup

List of split vector for each subgroup

call

Submitted arguments for the function (matched call)

fitobj

List of objects from subgroup estimation

References

Andersen, E. B. (1973). A goodness of fit test for the Rasch model. Psychometrika, 38(1), 123-140.

Fischer, G. H., & Molenaar, I. W. (Eds.). (2012). Rasch models: Foundations, recent developments, and applications. Springer Science & Business Media.

See also

Examples

# example for tmt_lrtest

# NOT RUN {
#############################################################################
# Example Rasch model and Likelihood Ratio Test
#############################################################################
dat <- tmt:::sim.rm(theta = 100,b = 10,seed = 1111)
dat.rm <- tmt_rm(dat = dat)
dat.lrt <- tmt_lrtest(dat.rm)
summary(dat.lrt)
# }