Function to create a template for the multistage design used in tmt

This function creates a template for the definition of multistage designs as required by the estimation function (in multistage design cases). The defines multistage design is then handed over to the functiontmt_mstdesign. Essentially, these are the modules, rules and path sections. In the formula-based notation, it is also possible to state additional conditions (constraints) that can be found in the data and are reflected in the multistage design.

tmt_msttemplate(formula = NULL, full = TRUE, eval = TRUE)

Arguments

formula

formula for the desired template of a multistage design. If formula is leaved empty, a matrix as MST design template is generated.

full

logical if the modules and rules sections should also be created

eval

logical should the text input be evaluated (e.g. 3:6 = c(3, 4, 5, 6))

Author

Jan Steinfeld

Examples

#############################################################################
# create simple template
#############################################################################
formula = "start(start) += S1(B1,B2,B3) += S2(B4,B5,B6,B7)"
tmt_msttemplate(formula, full = TRUE, eval = TRUE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> start( ) += B1( ) += B4( ) 
#> start( ) += B2( ) += B4( ) 
#> start( ) += B3( ) += B4( ) 
#> start( ) += B1( ) += B5( ) 
#> start( ) += B2( ) += B5( ) 
#> start( ) += B3( ) += B5( ) 
#> start( ) += B1( ) += B6( ) 
#> start( ) += B2( ) += B6( ) 
#> start( ) += B3( ) += B6( ) 
#> start( ) += B1( ) += B7( ) 
#> start( ) += B2( ) += B7( ) 
#> start( ) += B3( ) += B7( ) 
tmt_msttemplate(formula, full = TRUE, eval = FALSE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> start( ) += B1( ) += B4( ) 
#> start( ) += B2( ) += B4( ) 
#> start( ) += B3( ) += B4( ) 
#> start( ) += B1( ) += B5( ) 
#> start( ) += B2( ) += B5( ) 
#> start( ) += B3( ) += B5( ) 
#> start( ) += B1( ) += B6( ) 
#> start( ) += B2( ) += B6( ) 
#> start( ) += B3( ) += B6( ) 
#> start( ) += B1( ) += B7( ) 
#> start( ) += B2( ) += B7( ) 
#> start( ) += B3( ) += B7( ) 

#############################################################################
# create complex template
#############################################################################

formula = "nativ(no,yes) ~ education(low,medium,heigh) ~ 
            CBM(3:6) += S1(B1,B2,B3) += S2(B4,B5,B6,B7)"
tmt_msttemplate(formula, full = TRUE, eval = TRUE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3) += B3( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(4) += B3( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(5) += B3( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(6) += B3( ) += B7( ) 
tmt_msttemplate(formula, full = TRUE, eval = FALSE)
#> 
#> #-------------------- 
#> # modules 
#> #-------------------- 
#> #--------------------
#> # S1:
#> #-------------------- 
#> B1 =~  
#> B2 =~  
#> B3 =~  
#> #--------------------
#> # S2:
#> #-------------------- 
#> B4 =~  
#> B5 =~  
#> B6 =~  
#> B7 =~  
#>  
#> #-------------------- 
#> # rules 
#> #-------------------- 
#> Please fill here the specific rules
#>  
#>  
#> #-------------------- 
#> # mstdesign 
#> #-------------------- 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B4( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B5( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B6( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B1( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B2( ) += B7( ) 
#> nativ(no) ~ education(low) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(low) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(no) ~ education(medium) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(medium) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(no) ~ education(heigh) ~ CBM(3:6) += B3( ) += B7( ) 
#> nativ(yes) ~ education(heigh) ~ CBM(3:6) += B3( ) += B7( ) 

#############################################################################
# create template for the input as matrix
#############################################################################
tmt_msttemplate()
#>      mst              minSolved                                                
#> [1,] "paths"          "for each module the minimum raw score, seperated by ';'"
#> [2,] "^B4$-^B2$-^B1$" "0;0;0"                                                  
#>      maxSolved                                                
#> [1,] "for each module the maximum raw score, seperated by ';'"
#> [2,] "2;2;5"                                                  
#>      items                                                 
#> [1,] "for each module the specific items, seperated by ';'"
#> [2,] "i16,i17,i18,i19,i20;i6,i7,i8,i9,i10;i1,i2,i3,i4,i5"  
#>      minSolved_stage                                                   
#> [1,] "for each path the cumulative minimum raw score, seperated by ';'"
#> [2,] "0;0;0"                                                           
#>      maxSolved_stage                                                   
#> [1,] "for each path the cumulative maximum raw score, seperated by ';'"
#> [2,] "2;4;9"                                                           
#>      probability                                                                  
#> [1,] "for each path and module the probabilites (for deterministic routing use 1)"
#> [2,] "1,1,1,1,1,1;1,1,1,1,1,1;1,1,1,1,1,1"