Function for the Graphical Model Check

This function is a wrapper that processes the data of the likelihood ratio test for ggplot2. Items can be color coded and also excluded.

tmt_gmc(object, title = "graphical model check", xaxis = NULL,
  yaxis = NULL, lim = NULL, ellipse = FALSE, drop = NULL,
  alpha = 0.05, legendtitle = "split criteria", info = NULL)

Arguments

object
object of the function tmt_lrtest

title
of the plot

xaxis
description of the x-axis

yaxis
description of the y-axis

lim
of the plot

ellipse
should confidence-ellipse be plotted

drop
which items should be excluded from the plot

alpha
which alpha should be used for the ellipse

legendtitle
Title of the Legend

info
vector with further information for the Plot with names of submitted items

Examples

# NOT RUN {
#############################################################################
# Example of Graphical Model Check
#############################################################################
items <- seq(-3,3,length.out = 30)
names(items) <- paste0("i",1:30)
persons = 100
mean = 0
sd = 1
dat <- tmt:::sim.rm(theta = persons, beta = items,1111)
dat.rm <- tmt_rm(dat, optimization = "optim")
dat.lrt <- tmt_lrtest(dat.rm, split = "median", optimization = "optim")

info <- rep(c("group_a","group_b"),each = 15)
names(info) <- paste0("i",1:30)

drop <- c("i1","i18","i20","i100")

library(ggplot2)
tmt_gmc(object = dat.lrt,
        ellipse = TRUE,
        info = info,
        drop = drop,
        title = "graphical model check",
        alpha = 0.05,
        legendtitle = "split criteria")
# }