Function for the Graphical Model Check

This function performs a so-called graphical model check on the basis of the previously performed Likelihood Ratio Test [tmt::tmt_lrttest()]. The estimated item parameters of the two groups are plotted against each other. There is the possibility in this function to highlight items, to be excluded items from the plot, and to produce confidence-ellipses if desired.

tmt_gmc(
  object,
  title = "graphical model check",
  xaxis = NULL,
  yaxis = NULL,
  lim = NULL,
  ellipse = FALSE,
  drop = NULL,
  alpha = 0.05,
  legendtitle = "split criteria",
  info = NULL
)

Arguments

object

object of the function [tmt::tmt_lrttest()]

title

of the plot

xaxis

description of the x-axis

yaxis

description of the y-axis

lim

of the plot

ellipse

should confidence-ellipse be plotted

drop

which items should be excluded from the plot

alpha

which alpha should be used for the ellipse

legendtitle

Title of the Legend

info

vector with further information for the Plot with names of submitted items

Author

Jan Steinfeld

Examples

#############################################################################
# Example of Graphical Model Check
#############################################################################
items <- seq(-3,3,length.out = 16)
names(items) <- paste0("i",1:16)
persons = 500
dat <- tmt:::sim.rm(theta = persons, b = items, seed = 1234)
dat.rm <- tmt_rm(dat)
dat.lrt <- tmt_lrtest(dat.rm, split = "median")

info <- rep(c("group_a","group_b"),each = 8)
names(info) <- paste0("i",1:16)

drop <- c("i1","i18")

#library(ggplot2)
plot <- tmt_gmc(object = dat.lrt, 
  ellipse = TRUE, 
  info = info,
  drop = drop,
  title = "graphical model check",
  alpha = 0.05,
  legendtitle = "split criteria")